fix double to big decimal

+++ src/main/java/com/dto/material/MaterialCreateRequestDTO.java
@@ -4,11 +4,12 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public record MaterialCreateRequestDTO(String name,
                                        String description,
-                                       @NotNull @Positive Double price,
+                                       @NotNull @Positive BigDecimal price,
                                        @NotNull @PositiveOrZero Integer quantity,
                                        Set<String> categoryNames) {
 }
